+++ RELEASE-NOTES.md
@@ -6,6 +6,8 @@
 
 * Added support for the ternary conditional operator in expressions.  [Issue #47](https://github.com/mfenniak/rethinkdb-net/issues/47)
 
+* Runtime errors in expressions can be added with ReQLExpression.Error() in an expression tree.  [Issue #49](https://github.com/mfenniak/rethinkdb-net/issues/49)
+
 ### Breaking Changes
 
 * Added support for RethinkDB's binary format for byte[] conversion.  If byte[] was previously used by a client application, reading and writing it will begin using a different and incompatible data format.  Either migrate your data to the new format, or construct a datum converter that doesn't include the new BinaryDatumConverterFactory in it. [Issue #178](https://github.com/mfenniak/rethinkdb-net/issues/178)